Move `DeviceConfig`, ObservabilityConfig`, `SpeechToTextConfig` to their own files
